feat(editor): Implement go to start/end of file
Add `go_to_start_of_file` and `go_to_end_of_file` methods to the Editor.
Map `Alt/Option + <` to move the cursor to the start of the file.
Map `Alt/Option + >` to move the cursor to the end of the file.
New tests are included to verify these functionalities.