feat: Make undo debounce threshold configurable
Removes the `mock_instant` dependency and its associated feature flag.
Introduces a configurable `undo_debounce_threshold` in the `Editor` struct,
allowing the time for grouping consecutive undo actions to be set dynamically.
A new public method `set_undo_debounce_threshold` is added to facilitate this configuration.

Undo tests are updated to leverage this new functionality, setting a very
short debounce time for continuous input and then resetting it to zero
to simulate time passing beyond the threshold, thus eliminating the need
for the `mock_instant` library.